refactor: move strike observation logic into separate module as it became big enough
